structured 📕 and dialogue of add & cost added🚀

+++ lib/testing_phase/add_money_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class AddMoneyBloc extends Cubit<int> {
+class AddMoneyBloc extends Cubit<double> {
   AddMoneyBloc() : super(0);
 
   add(int data) {
